update tests and extract badge_component
